refactoring workers module into managers module
