+++ src/lib.rs
@@ -16,5 +16,5 @@ pub mod requests;
 pub mod udp_messages;
 pub mod pod_states;
 pub mod pod_data;
-pub mod workers;
+pub mod managers;
 pub mod error;
