+++ src/managers/mod.rs
@@ -6,11 +6,8 @@ pub mod worker_states {
 }
 pub mod messages;
 mod main_loop;
+mod udp;
+mod tcp;
 
-mod udp_worker;
-mod udp_manger;
-mod tcp_worker;
-mod tcp_manager;
-
-pub use udp_manger::UdpManager;
-pub use tcp_manager::TcpManager;
+pub use udp::UdpManager;
+pub use tcp::TcpManager;
