+++ src/managers/tcp/worker.rs
@@ -2,9 +2,9 @@ use crate::error::Error;
 use crate::requests;
 use crate::stream_utils;
 
-use super::worker_states::*;
-use super::messages::*;
-use super::main_loop::*;
+use super::super::worker_states::*;
+use super::super::messages::*;
+use super::super::main_loop::*;
 
 use std::io::prelude::*;
 use std::net::{
